+++ docs/CHANGELOG.rst
@@ -38,6 +38,7 @@ Changed
   ``bcm-workflow-chipseq``, ``bcm-workflow-wgbs``, ``dicty-align-reads``,
   ``dicty-etc``, ``affy`` and ``workflow-chip-seq``.
 - Expose additional parameters of ``bowtie2`` process
+- Support strandedness auto detection in ``qorts-qc`` process
 
 Fixed
 -----